rename Evaluator to Metrics and add more metrics
